--- src/main/kotlin/com/iplease/lib/bento/api/global/message/MessagePublishService.kt
@@ -1,5 +1,0 @@
-package com.iplease.lib.bento.api.global.message
-
-interface MessagePublishService {
-    fun publish(message: Message)
-}
