--- src/main/kotlin/com/iplease/lib/bento/api/global/message/MessageSubscribeService.kt
@@ -1,5 +1,0 @@
-package com.iplease.lib.bento.api.global.message
-
-interface MessageSubscribeService {
-    fun subscribe(message: Message)
-}
