--- src/main/kotlin/com/iplease/lib/bento/api/global/message/RouteMessageSubscribeService.kt
@@ -1,8 +1,0 @@
-package com.iplease.lib.bento.api.global.message
-
-import com.iplease.lib.bento.api.global.listener.MessageListener
-
-abstract class RouteMessageSubscribeService: MessageSubscribeService {
-    final override fun subscribe(message: Message) { getSubscribers().forEach { it.subscribe(message) } }
-    abstract fun getSubscribers(): Iterable<MessageListener>
-}
