+++ src/main/kotlin/com/iplease/lib/bento/api/global/subscribe/MessageSubscribeService.kt
@@ -1,0 +1,7 @@
+package com.iplease.lib.bento.api.global.subscribe
+
+import com.iplease.lib.bento.api.global.message.Message
+
+interface MessageSubscribeService {
+    fun subscribe(message: Message)
+}
